Added time to retrieve and fixed multiple compile warnings.

+++ Prototype/src/Server.h
@@ -47,7 +47,7 @@ namespace Prototype
 
 		//ServerTimeHandler timeHandler;
 		ServerGlobalObj serverGlobalObj;
-		float lastUpdateTime;
+		int lastUpdateTime;
 
 		//std::vector<ServerClient> clients;
